Search Module/makemytrip_automation/requirements.txt
selenium>=4.15.0
pytest>=7.4.0
pytest-html>=4.1.1
pytest-xdist>=3.3.1
webdriver-manager>=4.0.1
PyYAML>=6.0.1
Pillow>=10.0.0
python-dateutil>=2.8.2
undetected-chromedriver>=3.5.0
